refactor(util): extract Compose-free DateTimeUtil into util:core:public
`grocery/data/impl` (and other pure data-layer modules) only need the
`DateTimeUtil` interface, but it lived in `util:public`, which applies the
Compose plugin. That transitive Compose coupling blocks those modules from
targeting platforms without Compose Multiplatform (e.g. watchOS).

Move `DateTimeUtil` and the `LocalDate` helpers (`toInstantAtStartOfDay`,
`atStartOfDayIn`) into a new Compose-free module `client/util/core/public`,
keeping the `com.plusmobileapps.chefmate.util` package so no consumer imports
change. `util:public` re-exports the new module via `api`, and
`grocery/data/impl` now depends on `util:core:public` directly. Pure refactor,
no behavior change.
